Add axum-macros dependency and update axum features; refactor PrinterConfig port type and enhance OraclePrinter error handling

+++ momir_rs/src/config.rs
@@ -10,7 +10,7 @@ pub struct AppConfig {
 #[derive(Debug, Deserialize, Clone)]
 pub struct PrinterConfig {
     pub host: String,
-    pub port: i32,
+    pub port: u16,
 }
 
 #[derive(Debug, Deserialize, Clone)]
